Moved the component data to dimensions
