+++ app/src/main/java/com/example/hangoutz/ui/components/ErrorMessage.kt
@@ -7,15 +7,18 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import com.example.hangoutz.ui.theme.Error
+import com.example.hangoutz.utils.Dimensions
 
 
 @Composable
 fun ErrorMessage(errorMessage: String, modifier: Modifier = Modifier) {
     Row(
         modifier = modifier
-            .padding(start = 10.dp, end = 10.dp)
+            .padding(
+                start = Dimensions.ERROR_MESSAGE_PADDING_SMALL,
+                end = Dimensions.ERROR_MESSAGE_PADDING_SMALL
+            )
             .fillMaxWidth()
     ) {
         Text(
