+++ app/src/main/java/com/example/hangoutz/utils/Constants.kt
@@ -70,12 +70,13 @@ object Constants {
     //Validator strings
     const val AT_SIGN = "@"
     const val VALID_EMAIL = "@gmail.com"
-    const val CONTAINS_DIGIT = ".*\\\\d.*"
+    const val CONTAINS_DIGIT = ".*\\d.*"
 
     //RegisterViewModel
     const val MIN_NAME_LENGTH = 3
     const val MAX_NAME_LENGTH = 25
     const val MIN_PASSWORD_LENGTH = 8
+    const val TWO_HALVES = 2
     const val NAME_ERROR_MESSAGE = "Name must be 3–25 characters"
     const val EMAIL_FORMAT_ERROR_MESSAGE = "Email must be in correct format"
     const val EMAIL_DUPLICATE_ERROR_MESSAGE = "Email already in use"
